Move code to dedicated module
